refactor: separate filter functions
