+++ src/templates/components.ts
@@ -1,14 +1,14 @@
 abstract class Component {
-  protected container: HTMLElement;
+    protected container: HTMLElement;
 
-  constructor(tagName: string, className: string) {
-    this.container = document.createElement(tagName);
-    this.container.className = className;
-  }
+    constructor(tagName: string, className: string) {
+        this.container = document.createElement(tagName);
+        this.container.className = className;
+    }
 
-  render() {
-    return this.container;
-  }
+    render() {
+        return this.container;
+    }
 }
 
 export default Component;
